+++ src/drivers/lru-memory-cache-driver.ts
@@ -239,7 +239,7 @@ export class LRUMemoryCacheDriver
     }
 
     if (tags && tags.length > 0) {
-      await this.applyTags(parsedKey, tags);
+      await this.applyTags(key, tags);
     }
 
     this.log("cached", parsedKey);