+++ src/drivers/pg-cache-driver.ts
@@ -312,7 +312,7 @@ export class PgCacheDriver
       }
 
       if (tags && tags.length > 0) {
-        await this.applyTags(parsedKey, tags);
+        await this.applyTags(key, tags);
       }
 
       this.log("cached", parsedKey);
@@ -335,7 +335,7 @@ export class PgCacheDriver
       }
 
       if (tags && tags.length > 0) {
-        await this.applyTags(parsedKey, tags);
+        await this.applyTags(key, tags);
       }
 
       this.log("cached", parsedKey);
@@ -353,7 +353,7 @@ export class PgCacheDriver
     );
 
     if (tags && tags.length > 0) {
-      await this.applyTags(parsedKey, tags);
+      await this.applyTags(key, tags);
     }
 
     this.log("cached", parsedKey);